Create intervals.java
An interval is a range of numbers between two given numbers and includes all of the real numbers between those two numbers. For example, interval  includes all  such that .

Two intervals are said to be intersecting if they have any number in common. Given two intervals [a,b]  and [c,d] ,  report whether they are intersecting or not.

Eg: 1 5 4 9 
Output : YES
Explanation : 4 and 5 are common